requirements.txt
groq>=0.11.0
mss>=9.0.0
pillow>=10.0.0
google-genai>=1.0.0
playwright>=1.40.0
pymongo>=4.6.0
pyttsx3>=2.90.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
opencv-python>=4.8.0
pinecone>=8.1.2
pywhatkit>=0.6.0
pyautogui>=0.9.54
pycaw>=2023.3
comtypes>=1.2.0
The Pinecone library was recently renamed. 
1. Update `requirements.txt`: Change `pinecone-client` to just `pinecone`.
2. Ensure there are no other references to `pinecone-client` in the codebase.
